test_: split default messaging tests
